+++ meerkat-core/src/index.ts
@@ -4,9 +4,12 @@ export { detectApplyContextParamsToBaseSQL } from './context-params/context-para
 export * from './cube-measure-transformer/cube-measure-transformer';
 export * from './cube-to-duckdb/cube-filter-to-duckdb';
 export {
-  applyFilterParamsToBaseSQL, detectAllFilterParamsFromSQL, getFilterParamsAST
+  applyFilterParamsToBaseSQL,
+  detectAllFilterParamsFromSQL,
+  getFilterParamsAST,
 } from './filter-params/filter-params-ast';
 export { getWrappedBaseQueryWithProjections } from './get-wrapped-base-query-with-projections/get-wrapped-base-query-with-projections';
+export * from './joins/joins';
 export { FilterType } from './types/cube-types';
 export * from './types/cube-types/index';
 export * from './types/duckdb-serialization-types/index';
